+++ src/main/java/kr/co/skudeview/domain/Company.java
@@ -1,4 +1,4 @@
-package kr.co.skudeview.entity;
+package kr.co.skudeview.domain;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
@@ -12,23 +12,22 @@ import java.time.LocalDate;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class University extends BaseEntity{
+public class Company extends BaseEntity{
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "university_id")
+    @Column(name = "company_id")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
     private Member member_id;
 
-    @Column(name = "univ_name")
     @NotNull
-    private String univName;
+    @Column(name = "company_name")
+    private String companyName;
 
-    @Column(name = "major")
-    @NotNull
-    private String major;
+    @Column(name = "position")
+    private String position;
 
     @Column(name = "start_date")
     private LocalDate startDate;
@@ -39,11 +38,12 @@ public class University extends BaseEntity{
     @Column(name = "description")
     private String Description;
 
+
     @Builder
-    public University(Member member_id, String univName, String major, LocalDate startDate, LocalDate endDate, String description) {
-        this.member_id = member_id;
-        this.univName = univName;
-        this.major = major;
+    public Company(String companyName, String position, LocalDate startDate, LocalDate endDate, String description) {
+
+        this.companyName = companyName;
+        this.position = position;
         this.startDate = startDate;
         this.endDate = endDate;
         Description = description;
